refactor(iota-package-management): drop the iota-sdk dependency
`update_lock_file` and `update_lock_file_with_package_id` took a whole
`&WalletContext` but used it for only two values: the chain identifier and
the active environment's alias. That single edge put the JSON-RPC client
stack (`iota-json` -> `iota-json-rpc-types` -> `iota-json-rpc-api` ->
`iota-sdk`) in front of the Move build tooling and everything above it,
which also runs against the plan to phase out `iota-sdk`.

- take `chain_identifier` and `env_alias` as plain values in
  `update_lock_file_with_package_id`, matching what `set_package_id` and
  `resolve_published_id` already do; the function no longer needs to be
  async
- move `update_lock_file`, which extracts the package details from an
  `IotaTransactionBlockResponse`, into the CLI, its only caller; it
  resolves both values from the wallet context there
- drop `iota-sdk` and `iota-json-rpc-types` from
  `iota-package-management`, leaving `iota-protocol-config`,
  `iota-sdk-types`, `move-package` and `move-symbol-pool`

`iota-package-management` drops from build level 11 to 7 and
`iota-move-build` from 10 to 8, shortening the longest compilation chain
in the workspace from 18 levels to 17. No client code is left in the
Move build path.
