+++ crates/iota-package-management/Cargo.toml
@@ -15,9 +15,7 @@ anyhow.workspace = true
 thiserror.workspace = true
 
 # internal dependencies
-iota-json-rpc-types.workspace = true
 iota-protocol-config.workspace = true
-iota-sdk.workspace = true
 iota-sdk-types.workspace = true
 move-package.workspace = true
 move-symbol-pool.workspace = true
